--- services/ui-src/src/views/Home/index.module.scss
@@ -1,25 +1,0 @@
-.Home .lander {
-  padding: 80px 0;
-  text-align: center;
-}
-
-.Home .lander h1 {
-  font-family: "Open Sans", sans-serif;
-  font-weight: 600;
-}
-
-.Home .lander p {
-  color: #999;
-}
-
-.Home .amendments h4 {
-  font-family: "Open Sans", sans-serif;
-  font-weight: 600;
-  overflow: hidden;
-  line-height: 1.5;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-.Home .amendments p {
-  color: #666;
-}
