+++ services/ui-src/src/views/index.tsx
@@ -1,2 +1,4 @@
-export * from "./Home";
-export * from "./NotFound";
+import { Home } from "./Home/Home";
+import { NotFound } from "./NotFound/NotFound";
+
+export { Home, NotFound };
